feat: fully implement models for yandexgpt
