+++ storyteller-api/pkg/yandexgpt/message.go
@@ -1,0 +1,6 @@
+package yandexgpt
+
+type YandexGPTMessage struct {
+	Role yandexGPTRole `json:"role"`
+	Text string        `json:"text"`
+}
